Cambios en modulo de usuario

+++ client/src/components/EditArtist/index.js
@@ -6,14 +6,18 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import { history } from '../App';
 import './styles.css';
 import * as selectors from '../../reducers';
+import * as actionAlbums from '../../actions/albums';
+import * as actionArtists from '../../actions/artists';
+import * as actionTracks from '../../actions/tracks';
 import artistService from '../../services/artist';
+import albumService from '../../services/album';
 
 const artistModel = {
   artistid: null,
   artistname: "",
 };
 
-const EditArtist = ({ artist, artistAlbums, onSave, onDelete, selectColumn }) => {
+const EditArtist = ({ artist, artistAlbums, permissions, onSave, onDelete, selectColumn }) => {
   artist = {...artistModel, ...artist}
   document.body.style.backgroundColor = '#434343';
   const [nameInput, changeNameInput] = useState(artist.artistname);
@@ -21,7 +25,7 @@ const EditArtist = ({ artist, artistAlbums, onSave, onDelete, selectColumn }) =>
     <div className="edit-artist">
       <div className='edit-artist-content'>
         <div className='edit-artist-title'>
-          {(artist.artistid==null) ? 'Crear artista' : "Editar artista"}
+          {"Artista"}
         </div>
         <Table className="edit-artist-table" borderless>
           <thead>
@@ -34,48 +38,66 @@ const EditArtist = ({ artist, artistAlbums, onSave, onDelete, selectColumn }) =>
               <td>
                 <input className="edit-artist-input"
                   type="text"
-                  placeholder="Ingresa el nombre de la canción"
+                  placeholder="Ingresa el nombre de la artista"
                   value={nameInput}
                   onChange={e => changeNameInput(e.target.value)}
                 />
               </td>
               <td>
-                <button type="submit" className="edit-artist-button-save" onClick={() => 
-                  onSave({
-                    artistid: artist.artistid,
-                    artistname: nameInput,
-                  })
-                }>
-                  {(artist.artistid==null) ? 'Crear' : "Editar"}
+                {(artist.artistid==null && permissions.includes('Crear artista')) ?
+                  (<button type="submit" className="edit-artist-button-save" onClick={() => 
+                    onSave({
+                      artistid: artist.artistid,
+                      artistname: nameInput,
+                    })
+                  }>
+                    {'Crear'}
+                  </button>): null
+                }
+                {(artist.artistid!=null && permissions.includes('Editar artista')) ?
+                  (<button type="submit" className="edit-artist-button-save" onClick={() => 
+                    onSave({
+                      artistid: artist.artistid,
+                      artistname: nameInput,
+                    })
+                  }>
+                    {'Editar'}
+                  </button>): null
+                }                
+                {(artist.artistid!=null && permissions.includes('Eliminar artista')) ? 
+                  <button type="submit" className="edit-artist-button-delete" onClick={() => onDelete(artist, artistAlbums.length>0)}>
+                  {'Eliminar'}
+                </button> : 
+                  null
+                }
+                <button type="submit" className="edit-artist-button-delete w3-cyan" style={{marginLeft:'10px'}} onClick={() => history.goBack()}>
+                  {'Regresar'}
                 </button>
-              {(artist.artistid==null) 
-                ? <div/> 
-                : <button type="submit" className="edit-artist-button-delete" onClick={() => onDelete(artist)}>
-                    {'Eliminar'}
-                  </button>
-              }
               </td>
             </tr>
           </tbody>
         </Table>
-        <Table className='edit-artist-table-albums' size="sm" hover bordered>
-          <thead>
-            <tr>
-              <th>#</th>
-              <th>Album</th>
-            </tr>
-          </thead>
-          <tbody>
-            {artistAlbums.map((album, id) => 
-              (
-                <tr key={id} className={"table-light"} onClick={() => selectColumn(album)}>
-                  <th scope="row">{id+1}</th>
-                  <td>{album.albumname}</td>
-                </tr>
-              ))
-            }
-          </tbody>
-        </Table>
+        <div className="edit-artist-table-albums-container">
+          {'Álbums del artista:'}
+          <Table className='edit-artist-table-albums' size="sm" hover bordered>
+            <thead>
+              <tr>
+                <th>#</th>
+                <th>Álbum</th>
+              </tr>
+            </thead>
+            <tbody>
+              {artistAlbums.map((album, id) => 
+                (
+                  <tr key={id} className={"table-light"} onClick={() => selectColumn(album)}>
+                    <th scope="row">{id+1}</th>
+                    <td>{album.albumname}</td>
+                  </tr>
+                ))
+              }
+            </tbody>
+          </Table>
+        </div>
       </div>
     </div>
   );
@@ -86,20 +108,40 @@ export default connect(
   state => ({
     artist: selectors.getSelectedArtist(state),
     artistAlbums: selectors.getAlbums(state),
+    permissions: selectors.getLoggedUser(state).permissions,
   }),
   dispatch => ({
     onSave(artist) {
       if(artist.artistname!=="") {
       if(artist.artistid==null)
-        artistService.addArtist(artist).then(()=> history.push("/main/artistas"));
+        artistService.addArtist(artist).then(()=> history.push('/main/artistas'));
       if(artist.artistid!=null)
-        artistService.updateArtist(artist).then(()=> history.push("/main/artistas"));
+        artistService.updateArtist(artist).then(()=> history.push('/main/artistas'));
       }else{
         alert("Ingresa todos los campos para guardar.");
       };
     },
-    onDelete(artist) {
-      artistService.deleteArtist(artist).then(()=> history.push("/main/artistas"));
+    onDelete(artist, hasAlbums) {
+      if(hasAlbums) {
+        alert("Tiene que eliminar todos los álbums del artista para eliminar el artista.");
+      } else {  
+        artistService.deleteArtist(artist).then(()=> history.push('/main/artistas'));
+      }
+    },
+    selectColumn(album) {
+      dispatch(actionAlbums.selectAlbum(album));
+      dispatch(actionTracks.clearTracks());
+      albumService.getAlbumTracks(album).then(res=> {
+        const albumTracks = res;
+        albumTracks.map(track => dispatch(actionTracks.addTrack(track)));
+      });
+      //Se carga el DorpDown de artist
+      dispatch(actionArtists.clearArtists());
+      artistService.getArtistListAll().then(res=> {
+        const artistDropDown = res;
+        artistDropDown.map(artist => dispatch(actionArtists.addArtist(artist)));
+      });
+      history.push("/editarÁlbum");
     },
   }),
 )(EditArtist);
